refactoring & some small changes
