+++ src/pages/Test/Grammar/Grammar.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import './grammar.scss';
+import './Grammar.scss';
 
 const Grammar = () => {
   const grammarTasks = [
@@ -86,7 +86,8 @@ const Grammar = () => {
           isRight: false
         }
       ]
-    },{
+    },
+    {
       sentence: 'The best berries for me … strawberries.',
       options: [
         {
@@ -127,7 +128,8 @@ const Grammar = () => {
           isRight: false
         }
       ]
-    },{
+    },
+    {
       sentence: 'The best berries for me … strawberries.',
       options: [
         {
@@ -168,7 +170,50 @@ const Grammar = () => {
           isRight: false
         }
       ]
-    },{
+    },
+    {
+      sentence: 'The best berries for me … strawberries.',
+      options: [
+        {
+          option: 'is',
+          isRight: false
+        },
+        {
+          option: 'are',
+          isRight: true
+        },
+        {
+          option: 'be',
+          isRight: false
+        },
+        {
+          option: 'am',
+          isRight: false
+        }
+      ]
+    },
+    {
+      sentence: 'He … swim very well. He is the best swimmer in the team!',
+      options: [
+        {
+          option: 'can',
+          isRight: true
+        },
+        {
+          option: 'must',
+          isRight: false
+        },
+        {
+          option: 'should',
+          isRight: false
+        },
+        {
+          option: 'have to',
+          isRight: false
+        }
+      ]
+    },
+    {
       sentence: 'The best berries for me … strawberries.',
       options: [
         {
