+++ src/pages/Test/Test.scss
@@ -74,6 +74,7 @@
   font-size: 18px;
   color: #828282;
   padding-bottom: 5px;
+  text-align: center;
 }
 
 @media only screen and (max-width: 800px) {
